# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ICategory": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"company_id": {
						"type": "number",
						"format": "double"
					},
					"sort": {
						"type": "number",
						"format": "double"
					},
					"translations": {
						"items": {
							"properties": {
								"n": {
									"type": "string"
								},
								"t": {
									"type": "string"
								},
								"l": {
									"type": "string"
								}
							},
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITranslate": {
				"properties": {
					"l": {
						"type": "string"
					},
					"n": {
						"type": "string"
					},
					"t": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IVariantOrOptionForPosition": {
				"properties": {
					"n": {
						"type": "string"
					},
					"p": {
						"type": "number",
						"format": "double"
					},
					"q": {
						"type": "number",
						"format": "double"
					},
					"t": {
						"items": {
							"$ref": "#/components/schemas/ITranslate"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"EPriority": {
				"enum": [
					"FIRST",
					"SECOND"
				],
				"type": "string"
			},
			"IPositionForOrder": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"n": {
						"type": "string"
					},
					"p": {
						"type": "number",
						"format": "double"
					},
					"crt": {
						"type": "number",
						"format": "double"
					},
					"f": {
						"type": "number",
						"format": "double"
					},
					"t": {
						"items": {
							"$ref": "#/components/schemas/ITranslate"
						},
						"type": "array"
					},
					"v": {
						"$ref": "#/components/schemas/IVariantOrOptionForPosition"
					},
					"o": {
						"items": {
							"$ref": "#/components/schemas/IVariantOrOptionForPosition"
						},
						"type": "array"
					},
					"c": {
						"type": "string"
					},
					"i": {
						"type": "number",
						"format": "double"
					},
					"d": {
						"type": "number",
						"format": "double"
					},
					"cat": {
						"type": "number",
						"format": "double"
					},
					"pr": {
						"$ref": "#/components/schemas/EPriority"
					}
				},
				"required": [
					"cat"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOrder": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"n": {
						"type": "number",
						"format": "double"
					},
					"t": {
						"type": "number",
						"format": "double"
					},
					"p": {
						"items": {
							"$ref": "#/components/schemas/IPositionForOrder"
						},
						"type": "array"
					},
					"c": {
						"type": "string"
					},
					"crt": {
						"type": "number",
						"format": "double"
					},
					"d": {
						"type": "number",
						"format": "double"
					},
					"f": {
						"type": "number",
						"format": "double"
					},
					"m": {
						"type": "string"
					}
				},
				"required": [
					"t",
					"p",
					"c"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"company": {
						"type": "number",
						"format": "double"
					},
					"cat": {
						"type": "number",
						"format": "double"
					},
					"c": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"s": {
						"type": "number",
						"format": "double"
					},
					"n": {
						"type": "string"
					},
					"p": {
						"type": "number",
						"format": "double"
					},
					"d": {
						"type": "string"
					},
					"i": {
						"type": "string"
					},
					"a": {
						"type": "boolean"
					},
					"h": {
						"type": "boolean"
					},
					"v": {
						"items": {
							"properties": {
								"p": {
									"type": "number",
									"format": "double"
								},
								"n": {
									"type": "string"
								}
							},
							"type": "object"
						},
						"type": "array"
					},
					"o": {
						"items": {
							"properties": {
								"p": {
									"type": "number",
									"format": "double"
								},
								"n": {
									"type": "string"
								}
							},
							"type": "object"
						},
						"type": "array"
					},
					"t": {
						"items": {
							"properties": {
								"n": {
									"type": "string"
								},
								"t": {
									"type": "string"
								},
								"l": {
									"type": "string"
								}
							},
							"type": "object"
						},
						"type": "array"
					},
					"vt": {
						"items": {
							"items": {
								"properties": {
									"n": {
										"type": "string"
									},
									"t": {
										"type": "string"
									},
									"l": {
										"type": "string"
									}
								},
								"type": "object"
							},
							"type": "array"
						},
						"type": "array"
					},
					"ot": {
						"items": {
							"items": {
								"properties": {
									"n": {
										"type": "string"
									},
									"t": {
										"type": "string"
									},
									"l": {
										"type": "string"
									}
								},
								"type": "object"
							},
							"type": "array"
						},
						"type": "array"
					},
					"f": {
						"type": "string"
					},
					"fChanged": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "rests",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "sobogd ",
			"email": "sobogd@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/categories": {
			"post": {
				"operationId": "Categories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ICategory"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized request response"
					},
					"500": {
						"description": "Response with error"
					}
				},
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": []
			}
		},
...
		"/o/list-categories-with-positions": {
			"get": {
				"operationId": "List categories with positions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"i": {
												"items": {
													"$ref": "#/components/schemas/IItem"
												},
												"type": "array"
											},
											"c": {
												"type": "string"
											}
										},
										"required": [
											"i",
											"c"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized request response"
					},
					"500": {
						"description": "Response with error"
					}
				},
				"tags": [
					"O"
				],
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": []
			}
		},
		"/o/add-or-update-order": {
			"post": {
				"operationId": "Add or update order",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized request response"
					},
					"500": {
						"description": "Response with error"
					}
				},
				"tags": [
					"O"
				],
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IOrder"
							}
						}
					}
				}
			}
		},
		"/o/list-orders-for-table": {
			"post": {
				"operationId": "List orders for table",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IOrder"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized request response"
					},
					"500": {
						"description": "Response with error"
					}
				},
				"tags": [
					"O"
				],
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"tableId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"tableId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/o/order-by-number": {
			"post": {
				"operationId": "Order by number",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrder"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized request response"
					},
					"500": {
						"description": "Response with error"
					}
				},
				"tags": [
					"O"
				],
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"orderNumber": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"orderNumber"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/o/remove-order-by-number": {
			"post": {
				"operationId": "Remove order by number",
				"responses": {
					"204": {
						"description": "No content"
					},
					"401": {
						"description": "Unauthorized request response"
					},
					"500": {
						"description": "Response with error"
					}
				},
				"tags": [
					"O"
				],
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"orderNumber": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"orderNumber"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/o/finish-order-by-number": {
			"post": {
				"operationId": "Finish order by number",
				"responses": {
					"204": {
						"description": "No content"
					},
					"401": {
						"description": "Unauthorized request response"
					},
					"500": {
						"description": "Response with error"
					}
				},
				"tags": [
					"O"
				],
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"paymentMethod": {
										"type": "string"
									},
									"orderNumber": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"paymentMethod",
									"orderNumber"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/o/tables-with-orders": {
			"get": {
				"operationId": "Tables with orders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"f": {
												"type": "boolean"
											},
											"t": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"t"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized request response"
					},
					"500": {
						"description": "Response with error"
					}
				},
				"tags": [
					"O"
				],
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}